requirements.txt
aiosqlite==0.18.0
anyio==3.6.2
arrow==1.2.3
bbcode==1.1.0
binaryornot==0.4.4
cashews==5.0.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
chardet==5.1.0
charset-normalizer==3.0.1
chevron==0.14.0
click==8.1.3
contourpy==1.0.7
cookiecutter==1.7.3
cycler==0.11.0
distlib==0.3.6
fastapi==0.89.1
filelock==3.9.0
fonttools==4.38.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
importlib-metadata==6.0.0
Jinja2==3.1.2
jinja2-time==0.2.0
kiwisolver==1.4.4
loguru==0.6.0
MarkupSafe==2.1.2
matplotlib==3.6.3
msgpack==1.0.4
multidict==6.0.4
nb-cli==1.0.0
nonebot-adapter-onebot==2.2.0
nonebot-plugin-gocqhttp==0.6.4
nonebot-plugin-imageutils==0.1.13.5
nonebot2==2.0.0rc2
noneprompt==0.1.7
numpy==1.24.1
opencv-python-headless==4.7.0.68
packaging==23.0
Pillow==9.4.0
platformdirs==2.6.2
poyo==0.5.0
prompt-toolkit==3.0.36
psutil==5.9.4
py-cpuinfo==9.0.0
pydantic==1.10.4
pyfiglet==0.8.post1
pygtrie==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==7.0.0
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
text-unidecode==1.3
tomlkit==0.11.6
typing_extensions==4.4.0
ujson==5.7.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
virtualenv==20.17.1
watchfiles==0.18.1
wcwidth==0.2.6
websockets==10.4
yarl==1.8.2
zipp==3.11.0
